+++ src/lib/Core/Formats/Density/CIF.ts
@@ -12,12 +12,12 @@ namespace LiteMol.Core.Formats.Density.CIF {
 
         export function parse(block: Formats.CIF.DataBlock): ParserResult<Data> {
 
-            let info = block.getCategory('_density_info');
+            const info = block.getCategory('_density_info');
             if (!info) return ParserResult.error<Data>('_density_info category is missing.');
             if (!block.getCategory('_density_data')) return ParserResult.error<Data>('_density_data category is missing.');
 
             function getArray(name: string) {
-                let ret:number[] = [];
+                const ret:number[] = [];
                 for (let i = 0; i < 3; i++) {
                     ret[i] = info!.getColumn(`${name}[${i}]`).getFloat(0);
                 }
@@ -26,7 +26,7 @@ namespace LiteMol.Core.Formats.Density.CIF {
 
             function getNum(name: string) { return info!.getColumn(name).getFloat(0); }
 
-            let header = {
+            const header = {
                 name: info.getColumn('name').getString(0),
                 grid: getArray('grid'),
                 axisOrder: getArray('axis_order'),
@@ -39,46 +39,57 @@ namespace LiteMol.Core.Formats.Density.CIF {
                 spacegroupNumber: getNum('spacegroup_number') | 0,
             };
 
-            let alpha = (Math.PI / 180.0) * header.cellAngles[0],
+            const alpha = (Math.PI / 180.0) * header.cellAngles[0],
                 beta = (Math.PI / 180.0) * header.cellAngles[1],
                 gamma = (Math.PI / 180.0) * header.cellAngles[2];
 
-            let xScale = header.cellSize[0] / header.grid[0],
-                yScale = header.cellSize[1] / header.grid[1],
-                zScale = header.cellSize[2] / header.grid[2];
+            const xScale = header.cellSize[0],
+                yScale = header.cellSize[1],
+                zScale = header.cellSize[2];
 
-            let z1 = Math.cos(beta),
+            const z1 = Math.cos(beta),
                 z2 = (Math.cos(alpha) - Math.cos(beta) * Math.cos(gamma)) / Math.sin(gamma),
                 z3 = Math.sqrt(1.0 - z1 * z1 - z2 * z2);
 
-            let xAxis = [xScale, 0.0, 0.0],
+            const xAxis = [xScale, 0.0, 0.0],
                 yAxis = [Math.cos(gamma) * yScale, Math.sin(gamma) * yScale, 0.0],
                 zAxis = [z1 * zScale, z2 * zScale, z3 * zScale];
             
-            let indices = [0, 0, 0];
+            const indices = [0, 0, 0];
             indices[header.axisOrder[0]] = 0;
             indices[header.axisOrder[1]] = 1;
             indices[header.axisOrder[2]] = 2;
                         
-            let d = [header.origin[indices[0]], header.origin[indices[1]], header.origin[indices[2]]]
-            let origin = [
-                xAxis[0] * d[0] + yAxis[0] * d[1] + zAxis[0] * d[2],
-                                  yAxis[1] * d[1] + zAxis[1] * d[2],
-                                                    zAxis[2] * d[2]
-            ];
-            
-            let extent = [header.extent[indices[0]], header.extent[indices[1]], header.extent[indices[2]]];
+            const originGrid = [header.origin[indices[0]], header.origin[indices[1]], header.origin[indices[2]]]            
+            const extent = [header.extent[indices[0]], header.extent[indices[1]], header.extent[indices[2]]];
 
-            let rawData = readRawData1(block.getCategory('_density_data')!.getColumn('values'), extent, header.extent, indices, header.mean);            
-            let field = new Field3DZYX(<any>rawData.data, extent);                    
+            const rawData = readRawData1(block.getCategory('_density_data')!.getColumn('values'), extent, header.extent, indices, header.mean);            
+            const field = new Field3DZYX(<any>rawData.data, extent);                    
                                      
-            let data = Data.create(
-                header.cellSize, header.cellAngles, origin,
-                false, <any>void 0, field, extent,
-                { x: xAxis, y: yAxis, z: zAxis },
-                //[header.axisOrder[indices[0]], header.axisOrder[indices[1]], header.axisOrder[indices[2]]],
-                { min: rawData.min, max: rawData.max, mean: header.mean, sigma: header.sigma },
-                { spacegroupIndex: header.spacegroupNumber - 1, name: header.name });
+            // const data = Data.create(
+            //     header.cellSize, header.cellAngles, origin,
+            //     false, <any>void 0, field, extent,
+            //     { x: xAxis, y: yAxis, z: zAxis },
+            //     //[header.axisOrder[indices[0]], header.axisOrder[indices[1]], header.axisOrder[indices[2]]],
+            //     { min: rawData.min, max: rawData.max, mean: header.mean, sigma: header.sigma },
+            //     { spacegroupIndex: header.spacegroupNumber - 1, name: header.name });
+
+            const data: Data = {
+                spacegroup: {
+                    number: header.spacegroupNumber,
+                    size: header.cellSize,
+                    angles: header.cellAngles,
+                    basis: { x: xAxis, y: yAxis, z: zAxis }
+                },
+                box: {
+                    origin: [originGrid[0] / header.grid[0], originGrid[1] / header.grid[1], originGrid[2] / header.grid[2]],
+                    dimensions: [extent[0] / header.grid[0], extent[1] / header.grid[1], extent[2] / header.grid[2]],
+                    sampleCount: extent
+                },
+                data: field,
+                valuesInfo: { min: rawData.min, max: rawData.max, mean: header.mean, sigma: header.sigma },
+                attributes: { }
+            };
             
             return ParserResult.success(data);
         }